DTSPB-1238: initial changes
