+++ src/main/java/uk/gov/hmcts/probate/core/service/BusinessServiceImpl.java
@@ -3,9 +3,7 @@ package uk.gov.hmcts.probate.core.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
-import org.springframework.web.multipart.MultipartFile;
 import uk.gov.hmcts.probate.client.business.BusinessServiceApi;
-import uk.gov.hmcts.probate.client.business.BusinessServiceDocumentsApi;
 import uk.gov.hmcts.probate.client.submit.SubmitServiceApi;
 import uk.gov.hmcts.probate.core.service.mapper.ExecutorApplyingToInvitationMapper;
 import uk.gov.hmcts.probate.service.BusinessService;
@@ -30,7 +28,6 @@ import java.util.Optional;
 public class BusinessServiceImpl implements BusinessService {
 
     private final BusinessServiceApi businessServiceApi;
-    private final BusinessServiceDocumentsApi businessServiceDocumentsApi;
     private final SubmitServiceApi submitServiceApi;
     private final SecurityUtils securityUtils;
     private final ExecutorApplyingToInvitationMapper executorApplyingToInvitationMapper;
@@ -254,20 +251,6 @@ public class BusinessServiceImpl implements BusinessService {
             return businessServiceApi.pinNumber(phoneNumber, sessionId);
         }
     }
-
-    @Override
-    public List<String> uploadDocument(String authorizationToken, String userID, List<MultipartFile> files) {
-        if (files.isEmpty()) {
-            throw new IllegalArgumentException("There needs to be at least one file");
-        }
-        return businessServiceDocumentsApi.uploadDocument(userID, authorizationToken, files.get(0));
-    }
-
-    @Override
-    public String delete(String userID, String documentId) {
-        return businessServiceDocumentsApi.delete(userID, documentId);
-    }
-
 
     private ProbateCaseDetails getProbateCaseDetails(String caseId) {
         String serviceAuthorisation = securityUtils.getServiceAuthorisation();
