+++ src/main/java/uk/gov/hmcts/probate/service/BackOfficeService.java
@@ -1,9 +1,12 @@
 package uk.gov.hmcts.probate.service;
 
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.multipart.MultipartFile;
 import uk.gov.hmcts.probate.model.backoffice.GrantScheduleResponse;
 import uk.gov.hmcts.reform.probate.model.cases.CaseData;
 import uk.gov.hmcts.reform.probate.model.cases.ProbateCaseDetails;
+
+import java.util.List;
 
 public interface BackOfficeService {
 
@@ -22,4 +25,7 @@ public interface BackOfficeService {
     ResponseEntity<String> initiateExelaExtractDateRange(String fromDate, String toDate);
 
     ResponseEntity<String> initiateSmeeAndFordExtract(String fromDate, String toDate);
+
+    List<String> uploadDocument(String authorizationToken, List<MultipartFile> files);
+
 }
