+++ src/main/java/uk/gov/hmcts/probate/service/BusinessService.java
@@ -1,6 +1,5 @@
 package uk.gov.hmcts.probate.service;
 
-import org.springframework.web.multipart.MultipartFile;
 import uk.gov.hmcts.reform.probate.model.documents.BulkScanCoverSheet;
 import uk.gov.hmcts.reform.probate.model.documents.CheckAnswersSummary;
 import uk.gov.hmcts.reform.probate.model.documents.LegalDeclaration;
@@ -35,10 +34,6 @@ public interface BusinessService {
     List<Invitation> getAllInviteData(String formdataId);
 
     String getPinNumber(String phoneNumber, String sessionId, Boolean isBilingual);
-
-    List<String> uploadDocument(String authorizationToken, String userID, List<MultipartFile> files);
-
-    String delete(String userID, String documentId);
 
     List<Invitation> sendInvitations(List<Invitation> invitations, String sessionId, Boolean isBilingual);
 }
