+++ src/test/java/uk/gov/hmcts/probate/core/service/BackOfficeServiceImplTest.java
@@ -1,14 +1,18 @@
 package uk.gov.hmcts.probate.core.service;
 
+import com.google.common.collect.Lists;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
+import org.junit.jupiter.api.Assertions;
 import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.MockitoJUnitRunner;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.mock.web.MockMultipartFile;
 import uk.gov.hmcts.probate.client.backoffice.BackOfficeApi;
 import uk.gov.hmcts.probate.model.backoffice.BackOfficeCallbackRequest;
 import uk.gov.hmcts.probate.model.backoffice.BackOfficeCaveatData;
@@ -194,4 +198,26 @@ public class BackOfficeServiceImplTest {
         verify(backOfficeApi).initiateGrantAwaitingDocumentsNotification(eq("Bearer " + AUTHORIZATION),
             eq("Bearer " + SERVICE_AUTHORIZATION), eq(date));
     }
+
+    @Test
+    public void shouldUploadSuccessfully() {
+        MockMultipartFile file = new MockMultipartFile("file", "orig",
+            MediaType.IMAGE_PNG_VALUE, "bar".getBytes());
+        String authorizationToken = "AUTHTOKEN12345";
+        String serviceToken = "SERVICETOKEN67890";
+
+        when(securityUtils.getServiceAuthorisation()).thenReturn(serviceToken);
+        backOfficeService.uploadDocument(authorizationToken, Lists.newArrayList(file));
+
+        verify(backOfficeApi).uploadDocument(authorizationToken, serviceToken, file);
+    }
+
+    @Test()
+    public void shouldThrowExceptoinIfFilesAreEmpty() {
+        String authorizationToken = "AUTHTOKEN12345";
+        Assertions.assertThrows(IllegalArgumentException.class, () -> {
+            backOfficeService.uploadDocument(authorizationToken, Lists.newArrayList());
+        });
+    }
+
 }
